Fix copying TypeParams
There was a bug where the TypeParams weren't being properly copied but
rather continuing the use the source value.

The code:

```
cp := *x
if typeParams := typeparams.ForFuncType(x); typeParams != nil {
	*typeparams.ForFuncType(&cp) = *FieldList(typeParams)
}
```

cp.TypeParams which typeparams.ForFyncType returns is still
x.TypeParams. Then the `*typeparams.ForFuncType(&cp)` overwrites it a
copy thus modifying the original slice.

This issue causes a data race which showed up in golangci-lint and often
caused a panic.

The only way around this I found was to have copies of the function for
different versions of Go since the typeparam package doesn't have a way
to set the TypeParams field.
